迁移charPlus到单独模块，并添加单元测试
